# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Erica's Paint & Sip API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PaginationMeta {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
}

export interface Event {
  id: number;
  title: string;
  slug: string;
  date: string;
  durationMinutes?: number | null;
  location?: string | null;
  price?: number | null;
  capacity?: number | null;
  seatsAvailable?: number | null;
  description?: string | null;
  coverImageUrl?: string | null;
  isPublished: boolean;
  createdAt: string;
}

export interface CreateEventBody {
  title: string;
  date: string;
  durationMinutes?: number | null;
  location?: string | null;
  price?: number | null;
  capacity?: number | null;
  description?: string | null;
  coverImageUrl?: string | null;
  isPublished?: boolean;
}

export interface EventListResponse {
  data: Event[];
  meta: PaginationMeta;
}

export interface EventsSummary {
  totalEvents: number;
  upcomingEvents: number;
  totalBookings: number;
  totalRevenue: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  eventId: number;
  name: string;
  email: string;
  phone?: string | null;
  seats: number;
  paypalOrderId?: string | null;
  status: BookingStatus;
  createdAt: string;
}

export type CreateBookingBodyStatus =
  (typeof CreateBookingBodyStatus)[keyof typeof CreateBookingBodyStatus];

export const CreateBookingBodyStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface CreateBookingBody {
  eventId: number;
  name: string;
  email: string;
  phone?: string | null;
  seats: number;
  paypalOrderId?: string | null;
  status?: CreateBookingBodyStatus;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export interface BookingListResponse {
  data: Booking[];
  meta: PaginationMeta;
}

export interface PrivateInquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  preferredDate?: string | null;
  guestCount?: number | null;
  message: string;
  createdAt: string;
}

export interface CreatePrivateInquiryBody {
  name: string;
  email: string;
  phone: string;
  preferredDate?: string | null;
  guestCount?: number | null;
  message: string;
}

export interface PrivateInquiryListResponse {
  data: PrivateInquiry[];
  meta: PaginationMeta;
}

export interface GalleryPhoto {
  id: number;
  imageUrl: string;
  caption?: string | null;
  category?: string | null;
  sortOrder?: number | null;
  createdAt: string;
}

export interface CreateGalleryPhotoBody {
  imageUrl: string;
  caption?: string | null;
  category?: string | null;
  sortOrder?: number | null;
}

export interface GalleryPhotoListResponse {
  data: GalleryPhoto[];
  meta: PaginationMeta;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  publishedDate?: string | null;
  excerpt?: string | null;
  content?: string | null;
  coverImageUrl?: string | null;
  isPublished: boolean;
  createdAt: string;
}

export interface CreatePostBody {
  title: string;
  publishedDate?: string | null;
  excerpt?: string | null;
  content?: string | null;
  coverImageUrl?: string | null;
  isPublished?: boolean;
}

export interface PostListResponse {
  data: Post[];
  meta: PaginationMeta;
}

export interface UploadUrlRequest {
  /**
   * Original file name
   * @minLength 1
   */
  name: string;
  /**
   * File size in bytes
   * @minimum 1
   */
  size: number;
  /**
   * MIME type (e.g. image/jpeg)
   * @minLength 1
   */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload */
  uploadURL: string;
  /** Normalized object path (e.g. /objects/uploads/uuid). Store in DB. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface AdminMe {
  userId: string;
}

export interface NewsletterSubscriber {
  id: number;
  name?: string | null;
  email: string;
  subscribedAt: string;
}

export interface SubscribeNewsletterBody {
  name?: string | null;
  email: string;
}

export interface NewsletterSubscriberListResponse {
  data: NewsletterSubscriber[];
  meta: PaginationMeta;
}

export type ListEventsParams = {
  page?: number;
  limit?: number;
  /**
   * Only return future events
   */
  upcoming?: boolean;
};

export type GetUpcomingEventsParams = {
  limit?: number;
};

export type ListBookingsParams = {
  eventId?: number;
  status?: ListBookingsStatus;
  page?: number;
  limit?: number;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export type ListPrivateInquiriesParams = {
  page?: number;
  limit?: number;
};

export type ListGalleryPhotosParams = {
  category?: string;
  page?: number;
  limit?: number;
};

export type GetFeaturedGalleryPhotosParams = {
  limit?: number;
};

export type ListPostsParams = {
  page?: number;
  limit?: number;
};

export type GetRecentPostsParams = {
  limit?: number;
};

export type ListAdminEventsParams = {
  page?: number;
  limit?: number;
};

export type ListAdminPostsParams = {
  page?: number;
  limit?: number;
};

export type ListNewsletterSubscribersParams = {
  page?: number;
  limit?: number;
};