# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Erica's Paint & Sip API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin only — contains customer PII)
      parameters:
        - name: eventId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingListResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (called after PayPal success)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden — not in admin allowlist

  # ── Private Inquiries ────────────────────────────────────────────────────────
...
  /admin/me:
    get:
      tags: [admin]
      operationId: getAdminMe
      summary: Returns the signed-in user's Clerk ID (any authenticated user — no admin allowlist required; used for initial ADMIN_USER_IDS setup)
      responses:
        '200':
          description: Admin user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMe'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/events/{id}:
    get:
      tags: [admin]
      operationId: getAdminEvent
      summary: Get a single event by ID regardless of published status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Event record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/posts/{id}:
    get:
      tags: [admin]
      operationId: getAdminPost
      summary: Get a single post by ID regardless of published status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/events:
    get:
      tags: [admin]
      operationId: listAdminEvents
      summary: List all events including unpublished (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: All events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/posts:
    get:
      tags: [admin]
      operationId: listAdminPosts
      summary: List all posts including unpublished (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: All posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  # ── Newsletter ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ── Event ──────────────────────────────────────────────────────────────────
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        capacity:
          type: integer
          nullable: true
        seatsAvailable:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, date, isPublished, createdAt]
...
    EventListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        seats:
          type: integer
        paypalOrderId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, paid, cancelled]
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, name, email, seats, status, createdAt]
    CreateBookingBody:
      type: object
      properties:
        eventId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        seats:
          type: integer
        paypalOrderId:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, paid, cancelled]
          default: pending
      required: [eventId, name, email, seats]
    UpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, cancelled]
      required: [status]
    BookingListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]

    # ── Private Inquiry ────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        publishedDate:
          type: string
          format: date
          nullable: true
        excerpt:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, isPublished, createdAt]
...
    PostListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, meta]

    # ── Storage ────────────────────────────────────────────────────────────────
...
    AdminMe:
      type: object
      required: [userId]
      properties:
        userId:
          type: string

    # ── Newsletter ─────────────────────────────────────────────────────────────
...